test: minor improvements
