+++ src/sites/test-app/test-variant-browser-grid.js
@@ -1,0 +1,82 @@
+
+/**
+ * Copyright 2015-2023 OpenCB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {html, LitElement} from "lit";
+
+
+import "../../webcomponents/commons/forms/data-form.js";
+import {DATA_FORM_EXAMPLE} from "./conf/data-form.js";
+import {SAMPLE_DATA} from "./data/data-example.js";
+import {VARIANT_BROWSER_DATA} from "./data/variant-browser-data.js";
+
+class TestVariantBrowserGrid extends LitElement {
+
+    constructor() {
+        super();
+        this.#init();
+    }
+
+    createRenderRoot() {
+        return this;
+    }
+
+    static get properties() {
+        return {
+            opencgaSession: {
+                type: Object
+            },
+            config: {
+                type: Object
+            }
+        };
+    }
+
+    #init() {
+        const zip = new JSZip();
+        this._dataFormConfig = DATA_FORM_EXAMPLE;
+        this.configVariantGrid = {
+            pageSize: 10,
+            pageList: [10, 25, 50],
+            multiSelection: false,
+            showSelectCheckbox: false,
+            toolbar: {
+                // showNew: true,
+                showColumns: true,
+                showDownload: false,
+                showExport: false,
+                showSettings: false,
+                exportTabs: ["download", "link", "code"]
+                // columns list for the dropdown will be added in grid components based on settings.table.columns
+            },
+        };
+
+    }
+
+    render() {
+        return html`
+            <variant-browser-grid
+                .variants="${VARIANT_BROWSER_DATA}"
+                .opencgaSession="${this.opencgaSession}"
+                .config="${this.configVariantGrid}"
+                .populationFrequencies="${this.config.populationFrequencies}">
+            </variant-browser-grid>
+        `;
+    }
+
+}
+
+customElements.define("test-variant-browser-grid", TestVariantBrowserGrid);
